Added git ignore and cleaned up comments
